PWGDQ utility classes moved from AnalysisCore to Tasks/PWGDQ
