+++ Analysis/Core/src/AnalysisCoreLinkDef.h
@@ -16,8 +16,4 @@
 #pragma link C++ class TrackSelection + ;
 #pragma link C++ class TriggerAliases + ;
 
-#pragma link C++ class VarManager + ;
-#pragma link C++ class HistogramManager + ;
-#pragma link C++ class AnalysisCut + ;
-#pragma link C++ class AnalysisCompositeCut + ;
 #pragma link C++ class HFTrackIndexSkimsCreatorConfigs + ;
