+++ Analysis/Tasks/PWGDQ/AnalysisCompositeCut.cxx
@@ -8,7 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "AnalysisCore/AnalysisCompositeCut.h"
+#include "AnalysisCompositeCut.h"
 
 ClassImp(AnalysisCompositeCut)
