+++ Analysis/Tasks/PWGDQ/AnalysisCut.cxx
@@ -8,11 +8,7 @@
 // granted to it by virtue of its status as an Intergovernmental Organization
 // or submit itself to any jurisdiction.
 
-#include "AnalysisCore/AnalysisCut.h"
-
-#include <iostream>
-using std::cout;
-using std::endl;
+#include "AnalysisCut.h"
 
 ClassImp(AnalysisCut);
