+++ Analysis/Tasks/PWGDQ/PWGDQCoreLinkDef.h
@@ -11,3 +11,8 @@
 #pragma link off all globals;
 #pragma link off all classes;
 #pragma link off all functions;
+
+#pragma link C++ class VarManager + ;
+#pragma link C++ class HistogramManager + ;
+#pragma link C++ class AnalysisCut + ;
+#pragma link C++ class AnalysisCompositeCut + ;
