+++ Analysis/Tasks/PWGDQ/filterPP.cxx
@@ -19,10 +19,10 @@
 #include "AnalysisDataModel/Centrality.h"
 #include "AnalysisCore/TriggerAliases.h"
 #include "AnalysisDataModel/ReducedInfoTables.h"
-#include "AnalysisCore/VarManager.h"
-#include "AnalysisCore/HistogramManager.h"
-#include "AnalysisCore/AnalysisCut.h"
-#include "AnalysisCore/AnalysisCompositeCut.h"
+#include "VarManager.h"
+#include "HistogramManager.h"
+#include "AnalysisCut.h"
+#include "AnalysisCompositeCut.h"
 #include "CutsLibrary.h"
 #include "HistogramsLibrary.h"
 #include "AnalysisDataModel/PID/PIDResponse.h"
